% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_dim.R
\name{plotTree}
\alias{plotTree}
\title{Visualizing Euclidean Minimum Spanning Trees}
\usage{
plotTree(
  x,
  node.labels = "",
  node.col = "orange",
  node.size = 5,
  main = "MST plot",
  legend.cord = c(-1.2, 1.1)
)
}
\arguments{
\item{x}{input matrix, with rows as observations and columns as features}

\item{node.labels}{vector of labels for observations in x (vertices)}

\item{node.col}{vector of colors for the observations in x (vertices)}

\item{node.size}{numerical value to represent size of nodes in the plot}

\item{main}{title for the plot}

\item{legend.cord}{vector of the xy coordinates for the legend c(x,y)}
}
\value{
result plots a minimum spanning tree for input data x
}
\description{
Plots an Euclidean minimum spanning tree from given input data. If
 labels for the samples in the data are provided, a legend is also
  plotted.
}
