% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_dim.R
\name{separability}
\alias{separability}
\title{Separability of Labeled Data Points}
\usage{
separability(x, labels)
}
\arguments{
\item{x}{input data matrix, with rows as observations and columns
as features}

\item{labels}{a vector of labels for the observations. A label could
be a type of the observation e.g cell type in single-cell data}
}
\value{
A list with the following components:
\itemize{
 \item label_separability  A vector of separability scores for each
  of the label types. A high score denotes high separability
 \item overall_separability  Overall average separability score for
  all the labels
}
}
\description{
Computes homogeneity of labeled observations with multiple label
 types.
}
